+++ test/test.js
@@ -46,6 +46,19 @@ describe('etag(entity)', function () {
     })
   })
 
+  describe('when "entity" looks like a stats object', function () {
+    it('should generate a weak ETag', function () {
+      var fakeStat = {
+        atime: new Date('2014-09-01T14:52:07Z'),
+        ctime: new Date('2014-09-01T14:52:07Z'),
+        mtime: new Date('2014-09-01T14:52:07Z'),
+        ino: 0,
+        size: 3027
+      }
+      assert.equal(etag(fakeStat), 'W/"bd3-1182194534"')
+    })
+  })
+
   describe('with "weak" option', function () {
     describe('when "false"', function () {
       it('should generate a strong ETag for a string', function () {